+++ src/backend/configManager/defaultConfig.ts
