+++ src/backend/import/categoryCalculationScript-example.ts
