+++ src/ui/components/app/exporters/YnabExporter.vue
@@ -58,7 +58,7 @@
 
 <script lang="ts">
 import { setupExporterConfigForm } from '@/ui/components/app/exporters/exportersCommon';
-import { OutputVendorName } from '@/originalBudgetTrackingApp/commonTypes';
+import { OutputVendorName } from '@/backend/commonTypes';
 import YnabAccountMappingTable, { MappingTable } from '@/ui/components/app/exporters/YnabAccountMappingTable.vue';
 import { ref, defineComponent } from '@vue/composition-api';
