+++ src/ui/components/app/exporters/exportersCommon.ts
@@ -1,4 +1,4 @@
-import { OutputVendorName } from '@/originalBudgetTrackingApp/commonTypes';
+import { OutputVendorName } from '@/backend/commonTypes';
 import store from '@/ui/store';
 import { VForm } from '@/types/vuetify';
 import { computed, reactive, ref } from '@vue/composition-api';
