+++ src/ui/components/app/exporters/googleSheets/SheetsCombobox.vue
@@ -24,8 +24,8 @@
 import {
   computed, ref, onMounted, defineComponent, PropType
 } from '@vue/composition-api';
-import { getAllSpreadsheets, Spreadsheet } from '@/originalBudgetTrackingApp/export/outputVendors/googleSheets/googleSheetsInternalAPI';
-import { Credentials, createClient } from '@/originalBudgetTrackingApp/export/outputVendors/googleSheets/googleAuth';
+import { getAllSpreadsheets, Spreadsheet } from '@/backend/export/outputVendors/googleSheets/googleSheetsInternalAPI';
+import { Credentials, createClient } from '@/backend/export/outputVendors/googleSheets/googleAuth';
 import { required } from '@/ui/components/shared/formValidations';
 
 export default defineComponent({
